// src/store/actions/weatherActions.ts
import {ThunkAction} from "redux-thunk";
import {WeatherAction, WeatherData, WeatherError} from "../types";
import {RootState} from '..';
import {getData} from "../../api/apiCalls";
import {GET_WEATHER, SET_ERROR, SET_LOADING} from "../Constants";

export const getWeather = (city: string): ThunkAction<void, RootState, null, WeatherAction> => {
    return async dispatch => {
        try {
            const res = await getData(city)
            if (!res.ok) {
                const resData: WeatherError = await res.json();
                throw new Error(resData.message);
            }

            const resData: WeatherData = await res.json();
            dispatch({
                type: GET_WEATHER,
                payload: resData
            });
        } catch (err) {
            dispatch({
                type: SET_ERROR,
                payload: err.message
            });
        }
    }
}

export const setLoading = (): WeatherAction => {
    return {
        type: SET_LOADING
    }
}

export const setError = (): WeatherAction => {
    return {
        type: SET_ERROR,
        payload: ''
    }
}